new: add system tray icon
