+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
     "rapidfuzz==3.11.0",
     "pendulum>=3.0.0",
     "bjoern>=3.2.2",
+    "pystray>=0.19.5",
 ]
 
 [dependency-groups]
